make and use TopicsWidget
